+++ src/Microsoft.Azure.Connectors.Sdk/ConnectorClientBase.cs
@@ -8,8 +8,6 @@ using System.Text.Json.Serialization;
 using global::Azure.Core;
 using global::Azure.Core.Pipeline;
 using global::Azure.Identity;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
 
 namespace Microsoft.Azure.Connectors.Sdk
 {
@@ -37,7 +35,6 @@ namespace Microsoft.Azure.Connectors.Sdk
         };
 
         private readonly HttpPipeline _pipeline;
-        private readonly ILogger _logger;
         private readonly string _connectionRuntimeUrl;
         private bool _disposed;
 
@@ -59,7 +56,6 @@ namespace Microsoft.Azure.Connectors.Sdk
             options = ConnectorClientBase.ApplyBaseUri(options, connectionRuntimeUrl);
             credential ??= new DefaultAzureCredential();
 
-            this._logger = NullLogger.Instance;
             this._pipeline = HttpPipelineBuilder.Build(
                 options,
                 perRetryPolicies: new HttpPipelinePolicy[]
@@ -96,11 +92,6 @@ namespace Microsoft.Azure.Connectors.Sdk
         /// Gets the HTTP pipeline for making connector requests.
         /// </summary>
         protected HttpPipeline Pipeline => this._pipeline;
-
-        /// <summary>
-        /// Gets the logger instance.
-        /// </summary>
-        protected ILogger Logger => this._logger;
 
         /// <summary>
         /// Sends a connector API request and deserializes the JSON response.